Logic: applied the newly updated context and more
Many of the unecessary code because of the implementation of useReducer
are removed.

Other than that, the the MoveByOne component now give the user ability
to go backward one chapter if it is the last verse of the chapter. The
same thing for going forward one chapter is being implemented.